pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novel-factory"
version = "1.3.0"
description = "AI-powered novel chapter production factory"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
novelos = "novel_factory.cli:main"

[tool.setuptools.packages.find]
include = ["novel_factory*"]

[tool.setuptools.package-data]
novel_factory = ["db/schema/*.sql", "db/migrations/*.sql", "config/*.yaml", "*.yaml", "skill_packages/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
